test/ts4/requirements.txt
tonos-ts4==0.3.0
ton-client-py==1.20.0.0
